+++ sim-operations.h
@@ -1,6 +1,6 @@
 #ifndef SIM_OPERATIONS_H
 #define SIM_OPERATIONS_H
-/* genomicSimulationC v0.2.2.1 - last edit 11 Nov 2022 */
+/* genomicSimulationC v0.2.2.11 - last edit 8 June 2023 */
 
 #ifdef SIM_OPERATIONS
     #define RND_IMPLEMENTATION
@@ -123,7 +123,7 @@ typedef struct {
 */
 typedef struct {
 	int will_name_offspring; /**< A boolean: whether generated offspring should be given names. */
-	char* offspring_name_prefix; /**< If `will_name_offspring` is true, generated
+    const char* offspring_name_prefix; /**< If `will_name_offspring` is true, generated
                            * offspring are named [offspring_name_prefix][index]. */
 
 	int family_size; /**< The number of offspring to produce from each cross.*/
@@ -134,7 +134,7 @@ typedef struct {
                             * offspring of an anonymous individual (one without an ID)
                             * cannot identify that individual as their parent. */
 
-	char* filename_prefix; /**< A string used in save-as-you-go file names. */
+    const char* filename_prefix; /**< A string used in save-as-you-go file names. */
 	int will_save_pedigree_to_file; /**< A boolean. If true, the full/recursive
                             * pedigrees of every offspring generated in the cross
                             * are saved to "[filename_prefix}-pedigree.txt", even
@@ -707,7 +707,7 @@ void generate_clone(SimData* d, const char* parent_genome, char* output);
 
 int cross_random_individuals(SimData* d, const int from_group, const int n_crosses, const int cap, const GenOptions g);
 int cross_randomly_between(SimData*d, const int group1, const int group2, const int n_crosses, const int cap1, const int cap2, const GenOptions g);
-int cross_these_combinations(SimData* d, const int n_combinations, const int combinations[2][n_combinations],  const GenOptions g);
+int cross_these_combinations(SimData* d, const int n_combinations, const int* firstParents, const int* secondParents, const GenOptions g);
 int self_n_times(SimData* d, const int n, const int group, const GenOptions g);
 int make_doubled_haploids(SimData* d, const int group, const GenOptions g);
 int make_clones(SimData* d, const int group, const int inherit_names, const GenOptions g);